Fix CMakeLists to properly configure mutilsConfig.cmake
Using ${CMAKE_INSTALL_PREFIX} in mutilsConfig.cmake doesn't work because
that variable is not expanded when "make install" is executed, so the
mutilsConfig.cmake that gets installed to e.g. /usr/local/lib/cmake
still has the literal string ${CMAKE_INSTALL_PREFIX} in it. Then when
another Cmake project uses find_package(mutils), it loads
mutilsConfig.cmake and replaces ${CMAKE_INSTALL_PREFIX} with whatever
its current value is for that project, even if it's some other directory
like ~/.local/ and not the directory where mutils is installed. This
results in the confusing error message that mutils was found, but
libmutils.so could not be found, because Cmake is looking in the wrong
directory after expanding ${CMAKE_INSTALL_PREFIX} in mutils_LIBRARIES.

The fix is to use the configure_package_config_file command from
CMakePackageConfigHelpers, and the variable PACKAGE_CMAKE_INSTALL_LIBDIR
which gets expanded during make install. I copied these commands from
Derecho's CMakeLists.txt where they work correctly for installing the
package so that it can be used by another CMake project.